requirements.txt
Flask==3.0.3
openai==1.30.1
PyYAML==6.0.1
python-dotenv==1.0.1
pytest==8.2.0
pytest-flask==1.3.0
